fix(server): split cache.rs unit tests into a ratchet-exempt sibling file
apps/server/src/services/cache.rs had grown to 453 lines with its
#[cfg(test)] mod inline, tripping the module-size ratchet with no
allowlist row. Move the test module into cache_tests.rs (included via
#[path], same pattern as parquet_optimized.rs / parquet_optimized_tests.rs)
so it's exempt as a *_tests.rs sibling. cache.rs is now 208 lines.
